requirements.txt
acryl-datahub==1.2.0.9
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
appnope==0.1.4
asttokens==3.0.1
attrs==25.3.0
audioop-lts==0.2.2
audioread==3.1.0
avro==1.12.0
avro-gen3==0.7.16
beautifulsoup4==4.14.3
cached-property==2.0.1
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
click-default-group==1.2.4
click-spinner==0.1.10
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
datasets==4.5.0
debugpy==1.8.19
decorator==5.2.1
Deprecated==1.2.18
dill==0.4.0
distro==1.9.0
dnspython==2.8.0
docker==7.1.0
email-validator==2.3.0
executing==2.2.1
expandvars==1.1.2
faiss-cpu==1.13.2
fastapi==0.135.1
filelock==3.19.1
fonttools==4.61.1
frozenlist==1.7.0
fsspec==2025.9.0
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.1
humanfriendly==10.0
idna==3.10
ijson==3.4.0
ipykernel==7.1.0
ipython==9.9.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.3
jsonref==1.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
kiwisolver==1.4.9
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.46.0
MarkupSafe==3.0.2
matplotlib==3.10.8
matplotlib-inline==0.2.1
mixpanel==4.10.1
mpmath==1.3.0
msgpack==1.1.2
multidict==6.6.4
multiprocess==0.70.18
mypy_extensions==1.1.0
mysql-connector-python==9.6.0
nest-asyncio==1.6.0
networkx==3.5
numba==0.63.1
numpy==2.3.3
openai==2.15.0
packaging==25.0
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
pillow==12.1.1
platformdirs==4.5.1
pooch==1.8.2
progressbar2==4.5.0
prompt_toolkit==3.0.52
propcache==0.3.2
psutil==7.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==23.0.0
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
pyparsing==3.3.2
pypdf==6.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-utils==3.9.1
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.1.0
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
requests-file==2.1.0
rpds-py==0.27.1
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.13
safetensors==0.6.2
scikit-learn==1.8.0
scipy==1.17.0
sentry-sdk==2.38.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.8.3
soxr==1.0.0
stack-data==0.6.3
standard-aifc==3.13.0
standard-chunk==3.13.0
standard-sunau==3.13.0
starlette==0.52.1
sympy==1.14.0
tabulate==0.9.0
threadpoolctl==3.6.0
tokenizers==0.22.1
toml==0.10.2
torch==2.8.0
torchcodec==0.9.1
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
transformers==4.56.2
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.5.0
uvicorn==0.41.0
wcwidth==0.2.14
wheel==0.45.1
wrapt==1.17.3
xxhash==3.6.0
yarl==1.20.1
